scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
supabase==1.2.0
lxml==5.1.0
httpx==0.24.1
google-auth==2.27.0